Instrument ARM64 generated code to check the Marking Register.
Generate run-time code in the Optimizing compiler checking that
the Marking Register's value matches `self.tls32_.is.gc_marking`
in debug mode (on target; and on host with JIT, or with AOT when
compiling the core image). If a check fails, abort.

Generate similar run-time checks in JNI compiled code too.

Bug: 37707231
Test: m test-art-target
Test: Device boot test with `kEnableMarkingRegisterChecks` set to `true`
Change-Id: Ie9b322b22b3d26654a06821e1db71dbda3c43061