Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

Your CL comment isn't very descriptive. As far as I can see this is "only" wired up to the HandleUnexpectedSignal case? Which isn't on for the zygote? Is that intentional? If so, note that it's for testing.

If this is for testing only (which it weirdly implies because you're testing dump_on_stderr, which you likely should never see on device): you should have way better investigative opportunities with gdb...

So to me this CL needs *way* stronger arguing for why this code complexity / maintenance burden should be merged (after cleaning it up).

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1