Update patch set 5
Patch Set 5: Presubmit-Ready+1

> Is there anything here to make sure new memory is used for re-jit
 > methods? It didn't seem like it [at least not obviously so].
 > 
 > Reusing the same memory address as before will make it
 > significantly unlikely that any SIGILL will occur.

 > Is there anything here to make sure new memory is used for re-jit
 > methods? It didn't seem like it [at least not obviously so].
 > 
 > Reusing the same memory address as before will make it
 > significantly unlikely that any SIGILL will occur.

There are at least two failure modes that can occur:

1) SIGILL for fresh memory.
2) seemingly random crash because i$ includes stale cache lines from recycled memory.

 > Is there anything here to make sure new memory is used for re-jit
 > methods? It didn't seem like it [at least not obviously so].
 > 
 > Reusing the same memory address as before will make it
 > significantly unlikely that any SIGILL will occur.

That'd be work for a separate CL, but SIGILL isn't the only failure mode that might be expected if there are stale I$ lines.

For the SIGILL case there are 485 allocations in the test as-is that'd test that, not obvious, but. This is a tiny number for pushing the issue.

Patch-set: 5
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1