Fix HClinitCheck elimination in instruction builder.
To handle escaping instances of erroneous classes correctly,
we can omit the HClinitCheck only when using a class in the
static method of the very same class. Even for superclasses
we need to do the check. The new test exposes the cases
where we were previously diverging from the RI.

Also clean up the CompilerDriver by inlining one function
directly to the HInstructionBuild::IsInitialized(.) and
removing some related functions that are no longer used.

The size of the aosp_taimen-userdebug prebuilts:
  - before:
    arm/boot*.oat: 16891788
    arm64/boot*.oat: 19815520
    oat/arm64/services.odex: 20071624
  - after:
    arm/boot*.oat: 16949532 (+56.4KiB, +0.34%)
    arm64/boot*.oat: 19889752 (+72.5KiB, +0.37%)
    oat/arm64/services.odex: 20224328 (+149.1KiB, +0.76%)
with minor changes to other app prebuilts.

Note: Some of that could be reclaimed by reinstating the old
optimization for classes where no bytecode can be executed
during initialization (no <clinit> to execute in that class
or superclasses).

Test: 174-escaping-instance-of-bad-class
Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: testrunner.py --jvm -t 174
Test: Pixel 2 XL boots.
Test: testrunner.py --target --optimizing
Bug: 62478025
Change-Id: I41f026ea7fecc615c06e87f3b6cb847de0ede8a6