Remove manual libart.so run from test 018
Test 018 would always try to run twice, both with and without
libartd.so. This could cause some configurations to fail if they have
hard dependencies on one or the other (i.e. jvmti-stress runs).

Since both the debug and non-debug versions are tested on the build
servers we just make this run on whatever is requested by the tester.

Test: ./test/testrunner/testrunner.py --host --ndebug --debug -j40 -t 018
Change-Id: I917c693696a9154f500f5cae6325244e279ed6c6