ART: Allocate JIT code cache map as read-write only
Instead of requesting a chunk of memory as RWX immediately, only request
the large map covering both data and jit code cache as RW. The RemapAtEnd
will make the jit code cache tail RWX.

This works around a shortcoming of our linear-scan low-4GB map allocator.
When a process is not allowed to map executable pages, the allocator will
(very slowly) scan all 4GB. This cannot be solved locally, as the error
returned by the mmap call is the generic ENOMEM, not EACCES or EPERM.

Also use a unique_ptr to plug a leak on failure.

Bug: 36138508
Test: m
Test: m test-art-host
Test: Test SANITIZE_LITE=true build with wrap.system_server = asanwrapper
Change-Id: Id9e2e51bc8d7a29db99c406cd1e9ef6bbb8d444c