[WIP] ART: Change method lookup to conform to RI.
The method lookup for different invoke types methods was
previously widely different and didn't work well with the
dex cache method array where we have only a single slot for
each MethodId. The new behavior is to perform the same
lookup for all cases, distinguishing only between interface
and non-interface referencing class, and to further align
the behavior with the RI by giving preference to methods
declared in the superclass chain over "copied" methods which
come from interfaces. Where the JLS conflicts with JVMS, we
follow the JLS semantics.

The new lookup for class methods first searches the methods
declared in the superclass chain (ignoring "copied" methods)
and only then looks in the "copied" methods.

The new lookup for interface methods searches the interface
class and then the java.lang.Object class and then the
superinterfaces.

The new test 162-method-idx-clash exposes several cases
where we previously completely messed up due to the effects
of the DexCache, or where we were out of line with the RI.

TODO: Ignore private methods in superclasses (stick to JLS
      over JVMS). Add a test for this.

Test: art/test/run-test --host --jvm 162-method-resolution
Test: m test-art-host-gtest
Test: testrunner.py --host
Test: Nexus 6P boots.
Test: testrunner.py --target
Test: TODO maybe also cts?
Bug: 62855082
Bug: 30627598
Change-Id: If450c8cff2751369011d649c25d28a482a2c61a3