[WIP] ART: Change method lookup to conform to RI.
The lookup for direct and virtual methods was previously
different and didn't work well with the dex cache method
array where we have only a single slot for each MethodId.
The new behavior is to perform the same lookup for both
cases and to further align the behavior with the RI by
giving preference to methods declared in the superclass
chain over "copied" methods which come from interfaces.

The new test 162-method-idx-clash exposes a case where we
previously called the Base.foo() successfully when using
a MethodId referencing the Derived.foo() in contrast to
resolving to Derived.foo() and throwing IllegalAccessError
when performed from the Derived's defining dex file.
It also exposes the case where we previously preferred the
"copied" method from an interface over a static method
declared in a base class; the RI behavior is to prefer the
static method.

TODO: Lots of clean up.
TODO: Maybe some more tests.

Test: m test-art-host-gtest
Test: testrunner.py --host
Test: TODO maybe also cts and device tests?
Bug: 30627598
Change-Id: If450c8cff2751369011d649c25d28a482a2c61a3