[WIP] ART: Change method lookup to conform to RI.
The lookup for direct and virtual methods was previously
different and didn't work well with the dex cache method
array where we have only a single slot for each MethodId.
The new behavior is to perform the same lookup for both
cases and to further align the behavior with the RI by
giving preference to methods declared in the superclass
chain over "copied" methods which come from interfaces.

The new test 162-method-idx-clash exposes a case where we
previously called the Base.foo() successfully when using
a MethodId referencing the Derived.foo() in contrast to
resolving to Derived.foo() and throwing IllegalAccessError
when performed from the Derived's defining dex file.
It also exposes the case where we previously preferred the
"copied" method from an interface over a static method
declared in a base class; the RI behavior is to prefer the
static method.

TODO: Decide whether to ignore private methods in superclasses
      (diverge from JVM but in the spirit of javac).
TODO: Add interface method lookup paragraph to commit message.
TODO: Lots of clean up.
TODO: Some more tests.

Test: art/test/run-test --host --jvm 162-method-resolution
Test: m test-art-host-gtest
Test: testrunner.py --host
Test: Nexus 6P boots.
Test: testrunner.py --target
Test: TODO maybe also cts?
Bug: 62855082
Bug: 30627598
Change-Id: If450c8cff2751369011d649c25d28a482a2c61a3