[DISCUSSION] Method index clash.
Derived.test() calls the private Derived.foo() using
    invoke-direct {v0}, LDerived;->foo()V
while User.test() and User2.test() try to call the
Base.foo() using
    invoke-virtual {v0}, LDerived;->foo()V
which has an identical MethodId. The call from the class
User which is defined in a secondary dex file does so
successfully while the call from the class User2 which is
defined in the same dex file as Derived fails with
    java.lang.IllegalAccessError: Method 'void Derived.foo()'
    is inaccessible to class 'User2'
because the DexCache already contains resolved ArtMethod*
for that MethodId pointing to the private Derived.foo().

This is wrong as the resolution for invoke-virtual should
not look at direct methods at all.

Test: None; DO NOT SUBMIT
Bug: 30627598
Change-Id: If450c8cff2751369011d649c25d28a482a2c61a3