ART: Change method lookup to be closer to the RI.
The method lookup for different invoke types was previously
widely different and didn't work well with the dex cache
method array where we have only a single slot for each
MethodId. The new behavior is to perform the same lookup for
all cases, distinguishing only between interface and
non-interface referencing class, and to further align the
behavior with the RI. Where the JLS conflicts with JVMS, we
follow the JLS semantics.

The new lookup for class methods first searches the methods
declared in the superclass chain (ignoring "copied" methods)
and only then looks in the "copied" methods. If the search
in the supeclass chain finds a method that has not been
inherited (i.e. either a private method or a package-access
method where one of the classes in the chain does not belong
to the same package), we still search the copied methods as
there may actually be a method inherited from an interface.
This follows the JLS semantics where inherited methods are
included in the search (JLS 15.12.2.1) but conflicts with
the JVMS where the private or package-access method takes
precedence over methods inherited from interfaces.

Note that this search can find an accessible method that is
not inherited by the qualifying type, either for a package
access method when the referrer is in the same package but
the qualifying type is in another package, or for a private
method where the referrer is in the same class but the
qualifying type is actually a subclass. For the moment we
allow such calls and we shall consider whether to throw
an IncompatibleClassChangeError in this sutuation in future.

The new lookup for interface methods searches the interface
class, then the java.lang.Object class and then all the
superinterfaces. The search for the maximally-specific
non-abstract superinterface method is not yet implemented,
but the difference should be difficult to observe as the
usual subsequent call to FindVirtualMethodForInterface()
should yield the same result for any matching method.

The new test 162-method-idx-clash exposes several cases
where we previously completely messed up due to the effects
of the DexCache, or where we were out of line with the RI.
It also tests a case where the JVMS and JLS disagree and
we follow the JLS.

Test: art/test/run-test --host --jvm 162-method-resolution
Test: m test-art-host-gtest
Test: testrunner.py --host
Test: testrunner.py --host --interp-ac
Test: Nexus 6P boots.
Test: testrunner.py --target
Bug: 62855082
Bug: 30627598
Change-Id: If450c8cff2751369011d649c25d28a482a2c61a3

Change method lookup to follow JLS instead of JVMS.

If method lookup finds a method that is not inherited in the
superclass chain, we search the "copied" methods for another
method that is inherited from an interface and only return
the former when the search in "copied" methods fails. This
is a divergence from the RI but follows the semantics
implied by JLS (see 8.4.8 for inherited methods) and javac.