Update patch set 1
Patch Set 1:

> > I think what Andreas worries about, is individual engineers not
 > running them?
 > 
 > Yeah that's what he said to me.
 > 
 > I honestly don't think it's too huge a deal as long as the
 > buildbots are checking it and think that simplifying the testing
 > would be more useful.

Except downstream is *not* running libartd, and this is here exactly because we had breakages in the past in signalling code. I am still heavily against this.

(ndebug testing is not part of test-art-host or the default testrunner config, so yeah, *nobody* runs that by default.)

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>