Update patch set 1
Patch Set 1:

> > Patch Set 1: Code-Review-1
 > >
 > > Why was the previous change approved? This is braindead.
 > 
 > Can you clarify this a bit.
 > 
 > I made this change b/c iam@ asked me to make this target
 > consistent. I don't really care about it much tbh.

I am totally not happy with this naming. I think the previous decision to standardize on the old naming was wrong. We should change the layout of the build server.

But I won't block this change, so only -1, and someone else can approve.

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>