Update patch set 1
Patch Set 1:

I agree with you that the ideal place for such code is debuggerd but it is difficult to do so now. The functionality implemented in this patch is enough for us. So we've decided to apply this patch to Samsung code only. When jit code backtraces implementation is available in debuggerd we'll revert this patch and will use it.

Patch-set: 1
Reviewer: Gerrit User 1083288 <1083288@85c56323-6fa9-3386-8a01-6480fb634889>