Add support for class loader context in dex2oat
Initial dex2oat support for recongnizing the class loader context.

In order to correctly compile dex files which at runtime are loaded with
a non-trivial class loader chain we need to make dex2oat aware of the
precise runtime context.

This CL adds a way to pass a full class loader chain to dex2oat in order
to ensure that method resolution at compile time matches the resolution
at runtime.

The dex2oat option, '--class-loader-spec=<class-loader-spec>' accepts a
string specifying the intended runtime loading context for the compiled
dex files.

For example --classloader-spec=PCL[lib1.dex:lib2.dex];DLC[lib3.dex].

It describes how the class loader chain should be build in order to
ensure classes are resolved during dex2aot as they would be resolved at
runtime. This spec will be encoded in the oat file. If at runtime the
dex file will be loaded in a different context, the oat file will be
rejected.

The chain is interpreted in the natural 'parent order', meaning that
class loader 'i+1' will be the parent of class loader 'i'. The
compilation sources will be added to the classpath of the last class
loader. This allows the compiled dex files to be loaded at runtime in a
class loader that contains other dex files as well (e.g. shared
libraries).

The compiler will be tolerant if the source dex files specified with
--dex-file are found in the classpath. The source dex files will be
removed from the any class loader's classpath possibly resulting in
empty class loaders.

Test: m test-art-host
Bug: 38138251
Change-Id: I312aa12b5732288f3c1df4746b5775a32e0bfb04