Add support for class loader context in dex2oat
Initial dex2oat support for recongnizing the class loader context.

In order to correctly compile dex files which at runtime are loaded with
a non-trivial class loader chain we need to make dex2oat aware of the
precise runtime context.

This CL adds a way to pass a full class loader chain to dex2oat in order
to ensure that method resolution at compile time matches the resolution
at runtime.

Test: m test-art-host
Bug: 38138251
Change-Id: I312aa12b5732288f3c1df4746b5775a32e0bfb04