Fix boot classes init when compiling app image
Classes resident in boot library should not be initialized even if they
are not initialized and can be done while compiling app image since they
will later be written into app image and cause inconsistency.

Fixed by checking the location of classes in compiler driver when
initializing classes and is compiling app image.

The InitializeDependencies is advanced which makes TryInitialize become
the only entrypoint for app image, and since boot image doesn't need
this function, the behavior for compiling boot image should stay the same.

Bug: 62715577
Test: test-art-host -j64
Change-Id: I8e683e6b14f1402dba8ef5c4b0f5fa936a343d8c