Add support for generating boot image profile
Added three options:
--generate-boot-image-profile:
If either of these two options are passed in, profman creates a boot
profile.

--boot-image-clean-class-threshold=<value>
Specifies how many occurrences of a likely clean class are required
before a class is added to the profile.

--boot-image-class-threshold=<value>
Specify how many occurrences of a possibly dirty class are required
before a class is added to the profile.

Added unit test.

Test: test-art-host

Bug: 37966211

Change-Id: I8e12b0ec34dfa1d1bed0b51f342fffde09815348