Update patch set 2
Patch Set 2:

(1 comment)

Thanks for the insight, Tobias.

> I don't understand what this CL's purpose is; it seems to be doing three different things in the same CL, namely 1.) renaming a directory, 2.) fixing some tests, 3.) something about the bootclasspath?
> Can you be clearer in the CL description about what the bootclasspath change achieves?

Some of the tests do "import dalvik" or "import libcore" and those packages need to be on the javac classpath, otherwise the tests don't compile.

> What's the change in behavior - is it that hostside tests now exercise libcore code rather than the host toolchain's standard library?

It was always exercising the libcore code at runtime, it just wasn't compiling against libcore code without these changes.

> I'm wondering because the -bootclasspath command line option no longer exists in OpenJDK 9, so whatever this CL is doing item 3.) for will need to be redone differently for OpenJDK 9 support.

There has to be some way to specify where the java.base is, right? Through some environment variables or config files? Otherwise we have to edit the source code and change where it does the look up (and maybe that's the answer? we could have javac prebuilts instead of relying on the dev-system javac). 

This will be the same problem for all of Android since if we want to use @hide functions that aren't in upstream java.* packages, the compiler has to know about it.

The other packages (libcore, dalvik, etc) could probably be refactored into a separate classpath and provided to javac that way (without changing how it's looked up at runtime).

Patch-set: 2
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0