ART: Implement SuperblockCloner (1st part - Loop Unrolling).
SuperblockCloner provides a feature of cloning subgraphs in a
smart, high level way without fine grain manipulation with IR;
data flow and graph properties are resolved/adjusted automatically.
The clone transformation is defined by specifying a set of basic
blocks to copy and a set of rules how to treat edges, remap their
successors. By using this approach such optimizations as Branch
Target Expansion, Loop Peeling, Loop Unrolling can be implemented.

This CL provides interfaces and implementation of ControlFlow part
of the transformation; gtest is provided to test it.

Test: cloner_test.cc.
Change-Id: Ibeede38195376ca35f44ba9015491e50b3a5b87e