Update patch set 4
Patch Set 4:

> There has to be some way to specify where the java.base is, right?

Yes, --patch-module:

   http://openjdk.java.net/projects/jigsaw/quick-start#xoverride

My understanding is that --patch-module needs to be passed both to javac and java invocations.

I don't yet know whether this requires that source files for different OpenJDK 9 modules (eg. java.base) need to be split up into different compilation units (one per module) at compile time, e.g. a subdirectory for each module with the code for the classes for that module.

> This will be the same problem for all of Android since if we want to use @hide functions that aren't in upstream java.* packages, the compiler has to know about it.

Are you referring to the javadoc invocation (which is the only other invocation that I know of that uses the -bootclasspath argument)?

That invocation also needs to use --patch-module arguments for OpenJDK 9 because OpenJDK 9 javadoc doesn't know about -bootclasspath (bug 62049770). This is the biggest remaining issue that I know of blocking bug 38177295.

Note that this issue doesn't currently block us on javac because javac -source 1.8 -target 1.8 is a backwards compatible mode that doesn't know about modules, i.e. it allows compilation of classes that are in packages that belong to the java.base module without requiring any --patch-module arguments.

Patch-set: 4