Update patch set 1
Patch Set 1:

> I'm not sure we need this

I'm assuming "this" means "this patch", and not "noreturn".

The issue isn't so much that UNREACHABLE() is missing. It's more that clang is being very aggressive about assuming that code is always unreachable, when it's only actually unreachable in the given configuration.

If we marked the methods clang is warning about as `noreturn`, then we'd potentially need to add/remove some set of noreturn attributes every time we changed the project's configuration (e.g. if we flipped read barriers on, ...).

Ideally, there would be a `noreturn_if(kFoo || kBar || ...)` attribute that we could just tag on the functions clang is complaining about. Sadly, that doesn't seem to exist. :/

Patch-set: 1
Reviewer: Gerrit User 1096946 <1096946@85c56323-6fa9-3386-8a01-6480fb634889>