Fix skipping gtests.
Skipping of the target gtests was broken because parentheses were used
incorrectly. When ART_TEST_SKIP evaluates to 1, test is skipped and
ART_TEST_PASSED is not called which is correct. But ART_TEST_FAILED is not
called which is wrong because the value of the whole expression becomes 1
which is treated as error in make. The correct way is to make sure
ART_TEST_FAILED is called when the test is skipped. ART_TEST_FAILED handles
the skipped tests and evaluates as 0 so that make doesn't complain.

Test: Add a test to ART_TEST_KNOWN_BROKEN in Android.common_test.mk list and:
make test-art-target-gtest

Change-Id: I7d38723ab5f07fe461f645210c727ddc1f652a76