Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> What does --compiler-filter=verify do? I don't think we have go/ab or buildbot coverage for this, can we add one to go/ab?

@agampe says this filter level makes compiler not initialize classes, which is needed to reproduce the bug.

This is not isolated as an option, testcase 059 will try to do a test with core-verify.art inheriting it's original options, which is specified on the script named run. I don't know whether we need to configure the go/ab for this situation.

Patch-set: 5
Reviewer: Gerrit User 1167013 <1167013@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0