Don't do a read barrier in GetProfilingInfo.
ProfilingInfos are handled by the JIT, which treats declaring classes
of methods having these ProfilingInfo as weak roots. So it might be
possible the JIT requests a ProfilingInfo for a class that is going
to be unloaded (and that's ok, as the memory for the ProfilingInfo
is reclaimed by the JIT and is done before unloading the class).

bug: 62914384
Test: test.py
Change-Id: I84571786c1569782fb02d68257c4f70b195f27b6