Only look for optim. annotations in bootstrap class loader in Generic JNI.
When checking for an optimization annotation (FastNative or
CriticalNative) on a method in Generic JNI, do not resolve the
method's annotations' classes as a side effect -- instead, look them
up in the bootstrap class loader's resolved types. This is to prevent
exceptions from being thrown (during class resolution) in Generic JNI
transitions.

This change does not affect other cases of method optimization
annotation lookups (e.g. in non-Generic JNI), nor annotation lookup
rules in the context of reflection.

Test: art/test/testrunner/testrunner.py -t 656-annotation-lookup-generic-jni
Bug: 38454151
Bug: 34659969
Change-Id: Ie6b8b30b96a08baa629c449e3803a031515508d1