Update patch set 4
Patch Set 4:

(5 comments)

> This CL contains two things: unrolling and peeling; I've taken a
 > look into unrolling part and have put a couple of comments. However
 > I have a general comment: peeling is not turned on which leads to
 > having some portions of code being dead for the current moment. So
 > could you point me to the reason for having a single CL instead of
 > splitting it into two (implementing full support for peeling with
 > tests, etc)?


I wanted to combine all the loop duplicating logic in one CL, to show that we can now generate the peeling loop and unrolling loop with the same method. For good unrolling, we need to fill in the heuristics. For peeling, we need to access the address data.

So I was hoping that this CL would make the two follow up CLs simpler to read. I could postpone the peeling part until that later CL, or even introduce yet another one, but given the speed of reviews, I would really like to move forward with this.

Patch-set: 4
Reviewer: Gerrit User 1074526 <1074526@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0