Constrain time allowed for <clinit> in pre-init
Add a constraint for the time spent on <clinit> in aot compilers when
doing pre-initialization. The constraint is implemented by adding a
instrumentation listener which listen to DexPcMoved event, accumulating
the instructions executed and abort the transaction when it exceeds the
threashold.

The threashold for <clinit> can be specified by passing cmdline argument
to dex2oat, the default threashold is set to 100,000 per <clinit>.

A test is included which have a infinite loop in it's <clinit>, the
compiler fails to initialize the class whithout blocking too long time.

Test: make test-art-host -j64
Change-Id: I5268cfb157e45febacf98e4a5130503147e4efa3