Constrain time allowed for <clinit> in pre-init
Add a constraint for the time spent on <clinit> in aot compilers when
doing pre-initialization. The constraint is implemented by adding a
instrumentation listener which listen to DexPcMoved event, accumulating
the instructions executed and abort the transaction when it exceeds the
threshold.

The threshold for <clinit> can be specified by passing cmdline argument
to dex2oat, the default threshold is set to 100,000 per <clinit> to make
sure no class takes more than 1 second.

A test is included which has an infinite loop in its <clinit>, the
compiler fails to initialize the class whithout blocking too long time.

Test: make test-art-host -j64
Change-Id: I5268cfb157e45febacf98e4a5130503147e4efa3