Update patch set 8
Patch Set 8:

> > > I didn't find a natural way to split the CL. Do you care a lot
 > > > about it being split?
 > >
 > > There are two major parts: Adding CLCC, and then using it.
 > > Considering you gave CLCC a test, I think it is fine to have this
 > > as a standalone CL.
 > >
 > > (I'm just very averse to 1000-line CLs, because honestly, nobody
 > > reviews them that well...)
 > 
 > (Note that I'm OK if others think this is well-reviewed. :-) )

Oki, I'll wait for others to review as well. It's relative easy to split it as you suggested (one to introduce the context, the other to use the context). The usage is not that big (~200 lines of code including tests) and imo it gives a good sense on how the api is supposed to be used.

Patch-set: 8
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>