[RFC] ART: Collect owner contention stack on owner's unlock
To get precise owner stack traces (instead of where the owner is
at the point the contender reports contention), collect the stack
during Unlock and Wait. This requires the contender to store the
start of the wait into the monitor, and the owner to store the
stringified stack.

Collecting the stack on Unlock/Wait has two advantages. First, it
defers the work until the owner knows whether there might be
contention or not. Second, it works for the case where the owner was
originally holding a thin lock.

Mapping the unlock/wait location should usually give the same amount
of information as mapping the lock location. The stack dump will still
contain the information about the locked object, and except for
custom generated code or highly obfuscated code, locks and unlocks
can be easily mapped back to their scopes in the source.

Bug: 62353392
Test: m test-art-host
Change-Id: Iec5cfde995af693fe502ddef07b176151d2d826b