Update patch set 1
Patch Set 1:

> Richard WDYT?
 > 
 > Originally I was thinking to add a new field to the Size class, but
 > showing this new size separately doesn't give us much benefit, at
 > least for now, so I just extended your approach.

I like the general idea, but if we are going to use this approach more widely moving forward, this becomes a contract that platform (and app?) developers need to be aware of, and that Android Studio should be aware of. I would like to spend more time understanding the different use cases, formalizing a proposal, and getting feedback from platform and Android Studio developers first. This is something on my list of things to do for the next Android release.

How urgent is it for you to have this submitted in the short term? Do you have or have you made corresponding changes in Parcel required to make use of this? I'm not sure how I would like to move forward with this in the short term yet.

Patch-set: 1
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>