Change method lookup to follow JLS instead of JVMS.
Ignore private methods and package-access methods from
a different package when looking for a method in the
superclass chain. JVMS lookup rules say they are found
but the JLS says these are not inherited.

Remove regression test 601-method-access as this lookup
change makes it obsolete. The method would not be found
by the new lookup and as such we would not even enter
the access check where we removed the bogus DCHECK();
the DCHECK() would now hold in all situations.

Test: m test-art-host-gtest
Test: testrunner.py --host
Test: testrunner.py --host --interp-ac
Test: Nexus 6P boots.
Test: testrunner.py --target
Bug: 62855082
Change-Id: I250f51f10d6e0c29b5955677348be058fe1a35df