Update patch set 1
Patch Set 1:

Not urgent, but it's something nice to have. :)

I've made a parcel side change [http://ag/2480156] but it had a bigger performance penalty than I thought according to the microbenchmarks, so I'm tabling it for now.  (I'm hoping this is because how the microbenchmark works and because the code is too small and that makes the numbers skewed but I don't have time to make better benchmark.)

For now I need to apply my parcel patch locally when I need it, so I can just apply this change as well locally, so I can live without it merged in.

I thought of adding a map from Strings to integers to the Size class for an arbitrary number of "extra sizes", with a special naming convention (e.g. "mAhatNativeSizeXxx), but I felt like that was an overkill, since I don't think it's common for a class to have multiple native allocations for different purposes.  (Parcel does have two different kinds of allocation but the distinction isn't really useful for me.)


Another think I thought about was to make Parcel use NativeAllocationRegistry, but NAR doesn't seem to support native allocations with their size changing at runtime, so it didn't do it.  Also we don't want extra object allocations for Parcels.  (Parcel even has a object pool to reduce allocations.)

So again I can wait, but I felt like this is a simple change that would cover most usecases anyway.

Patch-set: 1
Reviewer: Gerrit User 1003801 <1003801@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0